+++ example_binary/src/main.rs
@@ -1,6 +1,7 @@
+use frost::Frost;
 use rove::{
     coordinator::start_server,
-    data_switch::{frost, DataSource, DataSwitch},
+    data_switch::{DataSource, DataSwitch},
     util::ListenerType,
 };
 use std::collections::HashMap;
@@ -8,7 +9,7 @@ use std::collections::HashMap;
 // TODO: use anyhow for error handling
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let data_switch = DataSwitch::new(HashMap::from([("frost", &frost::Frost as &dyn DataSource)]));
+    let data_switch = DataSwitch::new(HashMap::from([("frost", &Frost as &dyn DataSource)]));
 
     start_server(ListenerType::Addr("[::1]:1337".parse()?), data_switch).await
 }
