+++ frost/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "frost"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rove = { path = "../rove" }
+async-trait = "0.1.71"
+chrono = "0.4.26"
+chronoutil = "0.2.5"
+thiserror = "1.0.40"
+serde_json = "1.0.96"
+serde = { version = "1.0.164", features = ["derive"] }
+reqwest = { version = "0.11", features = ["json"] }
