+++ frost/src/duration.rs
